modified loader to deal with ppm files (not tested yet), some hard modifications on loader structure

+++ src/vector2d.cpp
@@ -5,7 +5,7 @@ Vector2D::Vector2D() {
     y = 0;
 }
 
-Vector2D::Vector2D(double x_, double y_) {
+Vector2D::Vector2D(float x_, float y_) {
     x = x_;
     y = y_;
 }
@@ -50,35 +50,35 @@ Vector2D Vector2D::operator-(const Vector2D& v) const {
     return t;
 }
 
-Vector2D & Vector2D::operator*=(const double a) {
+Vector2D & Vector2D::operator*=(const float a) {
     x *= a;
     y *= a;
     return *this;
 }
 
-Vector2D Vector2D::operator*(const double a)const {
+Vector2D Vector2D::operator*(const float a)const {
     Vector2D t = *this;
     t *= a;
     return t;
 }
 
-Vector2D operator*(const double a, const Vector2D & v) {
+Vector2D operator*(const float a, const Vector2D & v) {
     return Vector2D(v.x*a, v.y*a);
 }
 
-Vector2D& Vector2D::operator/=(const double a) {
+Vector2D& Vector2D::operator/=(const float a) {
     x /= a;
     y /= a;
     return *this;
 }
 
-Vector2D Vector2D::operator/(const double a)const {
+Vector2D Vector2D::operator/(const float a)const {
     Vector2D t = *this;
     t /= a;
     return t;
 }
 
-double Vector2D::length() const {
+float Vector2D::length() const {
     return sqrt(x*x+y*y);
 }
